streamlit/requirements.txt
streamlit==1.32.2
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.41.2
pyannote.audio==3.3.2
openai-whisper
pydub
huggingface_hub
numpy==1.26.4
pandas==2.2.2
accelerate
sentencepiece
bitsandbytes
jiwer
protobuf==3.20.3
